requirements.txt
Flask
Flask-Cors==4.0.1
Pillow
pikepdf>=8.0.0
gunicorn==22.0.0
zlib-state>=0.1.6 